# docs/swagger.yaml
basePath: /api/v1
definitions:
  requests.CreateBankAccountRequest:
    properties:
      account_name:
        type: string
      account_type:
        type: string
      balance:
        type: number
      category:
        type: string
      currency:
        type: string
      user_id:
        type: integer
    type: object
  requests.CreateCategoryRequest:
    properties:
      context:
        type: string
      context_type:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
...
  requests.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  responses.AccountResponse:
    properties:
      account_type:
        type: string
      balance:
        type: number
      category:
        type: string
      currency:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  responses.CategoryResponse:
    properties:
      context:
        type: string
      context_type:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  responses.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  responses.LoginResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      token:
        type: string
    type: object
  responses.TransactionResponse:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      category:
        type: string
      currency:
        type: string
      date:
        type: integer
      description:
        type: string
      id:
        type: integer
      transaction_type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Maybe Finance API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Maybe Finance API
  version: "1.0"
paths:
  /accounts:
    get:
      description: Retrieve all accounts
      parameters:
      - description: Group by account type
        enum:
        - true
        - false
        in: query
        name: group_by_account_type
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.AccountResponse'
            type: array
      summary: Get all accounts
      tags:
      - accounts
  /accounts/{id}:
    get:
      description: Retrieve an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.AccountResponse'
      summary: Get an account
      tags:
      - accounts
  /accounts/{id}/transactions:
    get:
      description: Retrieve an account's transactions
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TransactionResponse'
            type: array
      summary: Get an account's transactions
      tags:
      - accounts
  /accounts/{id}/transactions/recent:
    get:
      consumes:
      - application/json
      description: Get 5 recent transactions for an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TransactionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get 5 recent transactions for an account
      tags:
      - accounts
  /accounts/{id}/transactions/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload account transactions
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Transactions File
        in: formData
        name: file
        required: true
        type: file
      - description: Transaction Schema
        in: formData
        name: transaction_schema
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TransactionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Upload account transactions
      tags:
      - accounts
  /accounts/create:
    post:
      consumes:
      - application/json
      description: Create an account
      parameters:
      - description: Create Account Request
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/requests.CreateBankAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Create an account
      tags:
      - accounts
...
  /categories/{id}:
    get:
      description: Retrieve a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get a category
      tags:
      - categories
  /categories/create:
    post:
      consumes:
      - application/json
      description: Create a category
      parameters:
      - description: Create Category Request
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/requests.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Create a category
      tags:
      - categories
  /categories/types:
    get:
      description: Retrieve all category types
      parameters:
      - description: Context
        in: query
        name: context
        type: string
      - description: Context Type
        in: query
        name: context_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.CategoryResponse'
            type: array
      summary: Get all category types
      tags:
      - categories
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Login a user
      parameters:
      - description: Login Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.LoginResponse'
      summary: Login a user
      tags:
      - auth
swagger: "2.0"